pyproject.toml
[project]
name = "jewelry-design-platform-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofiles==24.1.0",
    "fastapi==0.115.0",
    "google-genai==1.9.0",
    "pillow==11.0.0",
    "pydantic[email]==2.9.0",
    "pydantic-settings==2.5.0",
    "python-dotenv==1.0.1",
    "python-multipart==0.0.12",
    "sqlalchemy==2.0.35",
    "uvicorn[standard]==0.32.0",
    "passlib>=1.7.4",
    "python-jose>=3.5.0",
    "bcrypt==4.3.0",
]
